Restructured models package.
