+++ src/models/alchemy/alchemy_event_types.rs
@@ -1,4 +1,4 @@
-use crate::models::alchemy_mined_transaction_data::AlchemyMinedTransactionData;
+use crate::models::alchemy::alchemy_mined_transaction_data::AlchemyMinedTransactionData;
 
 #[derive(Debug, Clone)]
 pub enum AlchemyEventTypes {
