+++ src/models/alchemy/mod.rs
@@ -1,0 +1,2 @@
+pub mod alchemy_event_types;
+pub mod alchemy_mined_transaction_data;
