+++ src/models/binance/mod.rs
@@ -1,0 +1,2 @@
+pub mod binance_event_types;
+pub mod binance_depth_data;
