+++ src/models/mod.rs
@@ -1,12 +1,9 @@
-pub mod alchemy_event_types;
-pub mod alchemy_mined_transaction_data;
+
+pub mod alchemy;
+pub mod binance;
+pub mod moralis;
+pub mod polygon;
 
 pub mod eth_transaction_types;
 
-pub mod moralis_erc20_token_price;
 
-pub mod polygon_crypto_aggregate_data;
-pub mod polygon_crypto_level2_book_data;
-pub mod polygon_crypto_trade_data;
-pub mod polygon_crypto_quote_data;
-pub mod polygon_event_types;
