+++ src/models/moralis/mod.rs
@@ -1,0 +1,1 @@
+pub mod moralis_erc20_token_price_response;
